Correct function definition in previous commit
Function argument definition did not match code.
Needed to change uint16_t to uint32_t

+++ src/slurmctld/reservation.c
@@ -158,7 +158,7 @@ static int  _select_nodes(resv_desc_msg_t *resv_desc_ptr,
 static int  _set_assoc_list(slurmctld_resv_t *resv_ptr);
 static void _set_cpu_cnt(slurmctld_resv_t *resv_ptr);
 static void _set_nodes_flags(slurmctld_resv_t *resv_ptr, time_t now,
-			     uint16_t flags);
+			     uint32_t flags);
 static int  _update_account_list(slurmctld_resv_t *resv_ptr,
 				 char *accounts);
 static int  _update_uid_list(slurmctld_resv_t *resv_ptr, char *users);
